\name{bundle_library}
\alias{bundle_library}
\title{Generates a path to a package's bundle library.}
\usage{
  bundle_library(pkg = ".")
}
\arguments{
  \item{pkg}{the path to the package}
}
\value{
  path to a package's bundle library
}
\description{
  Generates a path to a package's bundle library.
}

