\name{compare_versions}
\alias{compare_versions}
\title{Compares the requested version to the available version using the compare operator.}
\usage{
compare_versions(requested, compare, version)
}
\arguments{
  \item{requested}{the requested version}

  \item{compare}{the comparison operator}

  \item{version}{the available version}
}
\description{
Compares the requested version to the available version
using the compare operator.
}

