\name{construct_r_libs_user}
\alias{construct_r_libs_user}
\title{Constructs a new R_LIBS_USER setting using the current libraries and the new bundle library.}
\usage{
  construct_r_libs_user(bundle_path)
}
\arguments{
  \item{bundle_path}{the new bundle path}
}
\value{
  r_libs_user colon-separated libraries
}
\description{
  Constructs a new R_LIBS_USER setting using the current
  libraries and the new bundle library.
}

