\name{dependency_clauses}
\alias{dependency_clauses}
\title{Creates the `Depends:` clause by concatenating individual packages and adding their compare clauses.}
\usage{
  dependency_clauses(dependencies)
}
\arguments{
  \item{dependencies}{a data.frame with dependency package,
  compare, and version set.}
}
\description{
  Creates the `Depends:` clause by concatenating individual
  packages and adding their compare clauses.
}

