\name{depends_clause}
\alias{depends_clause}
\title{Creates the `Depends:` clause by concatenating individual packages and adding their compare clauses.}
\usage{
  depends_clause(dependencies)
}
\arguments{
  \item{dependencies}{a data.frame with dependency package,
  compare, and version set.}
}
\description{
  Creates the `Depends:` clause by concatenating individual
  packages and adding their compare clauses.
}

