\name{determine_version_to_install}
\alias{determine_version_to_install}
\title{Determines the version to install by comparing available versions to the required version and compare.}
\usage{
determine_version_to_install(available_versions, version, compare)
}
\arguments{
  \item{available_versions}{a vector of version identifiers
  corresponding to all versions of this package}

  \item{version}{the version requested}

  \item{compare}{the compare requested}
}
\description{
Determines the version to install by comparing available
versions to the required version and compare.
}

