\name{load_available_packages}
\alias{load_available_packages}
\title{Loads available packages from the given repository.}
\usage{
  load_available_packages(repos)
}
\arguments{
  \item{repos}{character vector, the base URLs of the
  repositories to use}
}
\value{
  data.frame of available packages
}
\description{
  Loads available packages from the given repository.
}

