\name{parse_sources}
\alias{parse_sources}
\title{Parses the sources information from DESCRIPTION file.}
\usage{
parse_sources(sources)
}
\arguments{
  \item{sources}{character vector of sources from the
  DESCRIPTION file}
}
\value{
data.frame with columns name, type, uri, and branch
}
\description{
Parses the sources information from DESCRIPTION file.
}

