\name{source_clauses}
\alias{source_clauses}
\title{Creates the `Sources:` clause by concatenating individual packages and adding their source clauses.}
\usage{
source_clauses(sources)
}
\arguments{
  \item{sources}{a data.frame with source package, type,
  uri, and optional branch set.}
}
\description{
Creates the `Sources:` clause by concatenating individual
packages and adding their source clauses.
}

