\name{update_current_environment}
\alias{update_current_environment}
\title{Updates the current environment.}
\usage{
update_current_environment(lib, r_libs_user)
}
\arguments{
  \item{lib}{the R library to add.}

  \item{r_libs_user}{the new value of R_LIBS_USER}
}
\description{
Updates the current environment.
}

