\name{update_renviron_file}
\alias{update_renviron_file}
\title{Updates a .Renviron file in the given path.}
\usage{
  update_renviron_file(path, r_libs_user)
}
\arguments{
  \item{path}{to the .Renviron file}

  \item{r_libs_user}{the new value of R_LIBS_USER}
}
\description{
  Updates a .Renviron file in the given path.
}

